Authentication implementation 0.1

+++ src/main/java/hu/flow/workoutTracker/Controller/ExerciseController.java
@@ -21,12 +21,12 @@ public class ExerciseController {
     }
 
     @GetMapping("/{id}")
-    public Exercise getExerciseById(@PathVariable int id){
+    public Exercise getExerciseById(@PathVariable long id){
         return exerciseService.getExerciseById(id);
     }
 
     @GetMapping("/byworkout/{workoutId}")
-    public List<Exercise> getAllExerciseByWorkout(@PathVariable int workoutId){
+    public List<Exercise> getAllExerciseByWorkout(@PathVariable long workoutId){
         return exerciseService.getAllExerciseByWorkout(workoutId);
     }
 
@@ -37,12 +37,12 @@ public class ExerciseController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Void> updateExcercise(@PathVariable int id, @RequestBody Exercise exercise){
+    public ResponseEntity<Void> updateExcercise(@PathVariable long id, @RequestBody Exercise exercise){
         return exerciseService.updateExercise(id, exercise);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteExercise(@PathVariable int id){
+    public ResponseEntity<Void> deleteExercise(@PathVariable long id){
         return exerciseService.deleteExercise(id);
     }
