+++ src/main/java/hu/flow/workoutTracker/Controller/WorkoutController.java
@@ -28,13 +28,13 @@ public class WorkoutController {
 
 
     @GetMapping("/currentuser/{userId}")
-    public List<Workout> getAllWorkoutByUser(@PathVariable int userId){
+    public List<Workout> getAllWorkoutByUser(@PathVariable long userId){
         return workoutService.getAllWorkoutByUser(userId);
     }
 
 
     @PostMapping("/currentuser/{userId}")
-    public ResponseEntity<Void> createWorkout(@PathVariable int userId, @RequestBody WorkoutRequestDTO workoutRequestDTO){
+    public ResponseEntity<Void> createWorkout(@PathVariable long userId, @RequestBody WorkoutRequestDTO workoutRequestDTO){
        return workoutService.createWorkout(userId, workoutRequestDTO);
     }
 
@@ -44,10 +44,9 @@ public class WorkoutController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteWorkout(@PathVariable int id){
+    public void deleteWorkout(@PathVariable long id){
         workoutService.deleteWorkout(id);
     }
-
 
 
 }
